--- apps/mobile/rust/Cargo.toml
@@ -1,39 +1,0 @@
-[package]
-name = "sd-core-mobile"
-version = "0.1.0"
-edition = "2021"
-rust-version = "1.64.0"
-
-[lib]
-crate-type = ["staticlib", "cdylib"] # staticlib for IOS and cdylib for Android
-
-[dependencies]
-once_cell = "1.15.0"
-sd-core = { path = "../../../core", features = [
-  "mobile",
-  "p2p",
-], default-features = false }
-rspc = { workspace = true }
-serde_json = "1.0.85"
-tokio = "1.21.2"
-openssl = { version = "0.10.42", features = [
-  "vendored",
-] } # Override features of transitive dependencies
-openssl-sys = { version = "0.9.76", features = [
-  "vendored",
-] } # Override features of transitive dependencies to support IOS Simulator on M1
-futures = "0.3.24"
-tracing = "0.1.37"
-
-[target.'cfg(target_os = "ios")'.dependencies]
-objc = "0.2.7"
-objc_id = "0.1.1"
-objc-foundation = "0.1.1"
-
-# This is `not(ios)` instead of `android` because of https://github.com/mozilla/rust-android-gradle/issues/93
-[target.'cfg(not(target_os = "ios"))'.dependencies]
-jni = "0.19.0"
-
-[target.'cfg(not(target_os = "ios"))'.features]
-default = ["sd-core/android"]
-
