+++ apps/mobile/rust/android/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "sd-core-android"
+version = "0.1.0"
+edition = "2021"
+rust-version = "1.64.0"
+
+[lib]
+# Android can use dynamic linking since all FFI is done via JNI
+crate-type = ["cdylib"]
+
+[dependencies]
+# FFI
+jni = "0.19.0"
+
+# Core mobile handling stuff
+sd-core-mobile = { path = "../mobile", features = ["android"] }
+
+# Other 
+tracing = "0.1.37"
