+++ apps/mobile/rust/mobile/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "sd-core-mobile"
+version = "0.1.0"
+edition = "2021"
+rust-version = "1.64.0"
+
+[features]
+android = ["sd-core/android"]
+ 
+[dependencies]
+once_cell = "1.15.0"
+sd-core = { path = "../../../../core", features = [
+  "mobile",
+  "p2p",
+], default-features = false }
+rspc.workspace= true
+serde_json = "1.0.85"
+tokio = "1.21.2"
+openssl = { version = "0.10.42", features = [
+  "vendored",
+] } # Override features of transitive dependencies
+openssl-sys = { version = "0.9.76", features = [
+  "vendored",
+] } # Override features of transitive dependencies to support IOS Simulator on M1
+futures = "0.3.24"
+tracing = "0.1.37"
